data-challenges/07-Data-Engineering/02-ML-Iteration/08-Cloud-Functions/CloudFunctionCode/requirements.txt
requests
bs4
google-cloud-storage
